+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### [Unreleased]
 
 - Add `ViewPrimaryKeyChecker` to enforce setting `primary_key` for models pointing to database views.
+- Improve `UniqueIndexChecker` and `MissingUniqueIndexChecker` to support partial indexes and uniqueness validators with `conditions` option.
 - Fix `UniqueIndexChecker` to skip partial indexes (indexes with a `WHERE` clause).
 - Fix `MissingUniqueIndexChecker` to skip uniqueness validators with a `conditions` option.